client_python/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "quantdata"
version = "1.3"
authors = [
    {name = "lzq", email = 'justgotpaid88@qq.com'}
]
description = 'quantdata client API'
requires-python = ">=3.9"
dependencies = ["numpy>=1.26.4", "pymongo", "pyyaml"]

[project.optional-dependencies]
tiledb = [ "tiledb" ]
duckdb = [ "duckdb" ]

[tool.setuptools]
packages = ["quantdata", "quantdata.analysis", "quantdata.databases"]

[tool.pytest.ini_options]
log_cli = false
log_file = "test.log"
log_file_level = "INFO"
log_file_format = "%(asctime)s %(levelname)s %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"